Homework New Year

+++ src/HWNY.java
@@ -15,17 +15,17 @@ public class HWNY {
         System.out.println("PART 2");
         System.out.println("--------------------------------------");
         //թվերը տրված են պատահական,կարաս հանգիստ փոխես
-        double a = 16;
-        double b = 44;
-        double c = 85;
-        double d = 65;
-        double e = 45;
-        double f = 33;
-        double i = (a + b + c) / 3;
-        double j = (d + e + f) / 3;
+        int a = 16;
+        int b = 44;
+        int c = 85;
+        int d = 65;
+        int e = 45;
+        int f = 33;
+        double i = (double) (a + b + c) / 3;
+        double j = (double) (d + e + f) / 3;
         System.out.println(i);
         System.out.println(j);
-        System.out.println("k=" + (int) (i + j)  );
+        System.out.println("k=" + (int)(i + j));
 
         System.out.println("Happy գալիք  New Year Դավ ջան,քեզ ամենալավ բաները");
     }
